requirements.txt
requests==2.31.0
python-dotenv==1.0.0
resend==0.7.2
beautifulsoup4==4.12.2